Remove eth from project
