+++ networks/__init__.py
@@ -1,2 +1,1 @@
 from .duc import DucMaker
-from .eth import EthMaker
