--- networks/eth/__init__.py
@@ -1,1 +1,0 @@
-from .starter import EthMaker
